Refining Move Project Registry
The foundation plans to publish a registry of projects which Aptos is interested in and which could receive grants. This PR refines the description of projects for Move and Developer platform in this registry, namely:

- DApp for Move package catalog
- Move linter
- Move formatter
- Move decompiler
- Move specification tester

Note that this speficially does _not_ contain (a) VS Code support (already has applications) (b) Hardhat for Move (currently planned to be done by Labs themselves). Notice that in general, we can add more projects and/or remove some and do them ourselves if no grant has been assigned yet.